Deprecate old output functions, move new_ferc1.py into ferc1.py
