+++ src/pudl/output/ferc1.py
@@ -1,89 +1,45 @@
-"""Functions for pulling FERC Form 1 data out of the PUDL DB."""
+"""A collection of denormalized FERC assets and helper functions."""
 import numpy as np
 import pandas as pd
-import sqlalchemy as sa
+from dagster import Field, asset
 
 import pudl
-from pudl.metadata.fields import apply_pudl_dtypes
 
 logger = pudl.logging_helpers.get_logger(__name__)
 
 
-def read_table_with_start_end_dates(
-    table_name: str,
-    start_date: pd.Timestamp,
-    end_date: pd.Timestamp,
-    pudl_engine: sa.engine.Engine,
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_plants_utilities_ferc1(
+    plants_ferc1: pd.DataFrame,
+    utilities_ferc1: pd.DataFrame,
 ) -> pd.DataFrame:
-    """Read a FERC1 table with a start and end date restriction.
-
-    Note: This is a FERC1 function instead of a general function because FERC1 uses
-    ``report_year`` instead of the more standard ``report_date``.
-
-    Args:
-        table_name: Name of FERC1 table.
-        start_date: Beginning date for data to pull from the PUDL DB.
-        end_date: End date for data to pull from the PUDL DB.
-        pudl_engine: : A connection engine for the PUDL DB.
-    """
-    pt = pudl.output.pudltabl.get_table_meta(pudl_engine)
-    table = pt[table_name]
-    table_select = (
-        sa.sql.select(table)
-        .where(table.columns.report_year >= f"{start_date.year}")
-        .where(table.columns.report_year <= f"{end_date.year}")
-    )
-    return pd.read_sql(table_select, pudl_engine)
+    """A denormalized table containing FERC plant and utility names and IDs."""
+    return pd.merge(plants_ferc1, utilities_ferc1, on="utility_id_ferc1")
 
 
-def plants_utils_ferc1(pudl_engine):
-    """Build a dataframe of useful FERC Plant & Utility information.
-
-    Args:
-        pudl_engine (sqlalchemy.engine.Engine): Engine for connecting to the
-            PUDL database.
-
-    Returns:
-        pandas.DataFrame: A DataFrame containing useful FERC Form 1 Plant and
-        Utility information.
-    """
-    logger.warning(
-        "pudl.output.ferc1.plants_utils_ferc1() will be deprecated in a future version of PUDL."
-        " In the future, call the PudlTabl.pu_ferc1() method or pull the plants_utils_ferc1 table"
-        "directly from the pudl.sqlite database."
-    )
-    pu_df = pd.merge(
-        pd.read_sql("plants_ferc1", pudl_engine),
-        pd.read_sql("utilities_ferc1", pudl_engine),
-        on="utility_id_ferc1",
-    )
-    return pu_df
-
-
-def plants_steam_ferc1(pudl_engine, start_date, end_date):
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_plants_steam_ferc1(
+    denorm_plants_utilities_ferc1: pd.DataFrame, plants_steam_ferc1: pd.DataFrame
+) -> pd.DataFrame:
     """Select and joins some useful fields from the FERC Form 1 steam table.
 
     Select the FERC Form 1 steam plant table entries, add in the reporting
     utility's name, and the PUDL ID for the plant and utility for readability
     and integration with other tables that have PUDL IDs.
-
     Also calculates ``capacity_factor`` (based on ``net_generation_mwh`` &
     ``capacity_mw``)
 
     Args:
-        pudl_engine (sqlalchemy.engine.Engine): Engine for connecting to the
-            PUDL database.
+        denorm_plants_utilities_ferc1: Denormalized dataframe of FERC Form 1 plants and
+            utilities data.
+        plants_steam_ferc1: The normalized FERC Form 1 steam table.
 
     Returns:
-        pandas.DataFrame: A DataFrame containing useful fields from the FERC
-        Form 1 steam table.
+        A DataFrame containing useful fields from the FERC Form 1 steam table.
     """
     steam_df = (
-        read_table_with_start_end_dates(
-            "plants_steam_ferc1", start_date, end_date, pudl_engine
-        )
-        .merge(
-            plants_utils_ferc1(pudl_engine),
+        plants_steam_ferc1.merge(
+            denorm_plants_utilities_ferc1,
             on=["utility_id_ferc1", "plant_name_ferc1"],
             how="left",
         )
@@ -113,6 +69,678 @@ def plants_steam_ferc1(pudl_engine, start_date, end_date):
         )
     )
     return steam_df
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_plants_small_ferc1(
+    plants_small_ferc1: pd.DataFrame, denorm_plants_utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe related to the FERC Form 1 small plants."""
+    plants_small_df = (
+        plants_small_ferc1.merge(
+            denorm_plants_utilities_ferc1,
+            on=["utility_id_ferc1", "plant_name_ferc1"],
+            how="left",
+        )
+        .assign(
+            opex_total=lambda x: (
+                x[["opex_fuel", "opex_maintenance", "opex_operations"]]
+                .fillna(0)
+                .sum(axis=1)
+            ),
+            opex_total_nonfuel=lambda x: (x.opex_total - x.opex_fuel.fillna(0)),
+        )
+        .pipe(
+            pudl.helpers.organize_cols,
+            [
+                "report_year",
+                "utility_id_ferc1",
+                "utility_id_pudl",
+                "utility_name_ferc1",
+                "plant_id_pudl",
+                "plant_name_ferc1",
+                "record_id",
+            ],
+        )
+    )
+
+    return plants_small_df
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_plants_hydro_ferc1(
+    plants_hydro_ferc1: pd.DataFrame, denorm_plants_utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe related to the FERC Form 1 hydro plants."""
+    plants_hydro_df = (
+        plants_hydro_ferc1.merge(
+            denorm_plants_utilities_ferc1,
+            on=["utility_id_ferc1", "plant_name_ferc1"],
+            how="left",
+        )
+        .assign(
+            capacity_factor=lambda x: (x.net_generation_mwh / (8760 * x.capacity_mw)),
+            opex_total_nonfuel=lambda x: x.opex_total,
+        )
+        .pipe(
+            pudl.helpers.organize_cols,
+            [
+                "report_year",
+                "utility_id_ferc1",
+                "utility_id_pudl",
+                "utility_name_ferc1",
+                "plant_name_ferc1",
+                "record_id",
+            ],
+        )
+    )
+    return plants_hydro_df
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_plants_pumped_storage_ferc1(
+    plants_pumped_storage_ferc1: pd.DataFrame,
+    denorm_plants_utilities_ferc1: pd.DataFrame,
+) -> pd.DataFrame:
+    """Pull a dataframe of FERC Form 1 Pumped Storage plant data."""
+    pumped_storage_df = (
+        plants_pumped_storage_ferc1.merge(
+            denorm_plants_utilities_ferc1,
+            on=["utility_id_ferc1", "plant_name_ferc1"],
+            how="left",
+        )
+        .assign(
+            capacity_factor=lambda x: x.net_generation_mwh / (8760 * x.capacity_mw),
+            opex_total_nonfuel=lambda x: x.opex_total,
+        )
+        .pipe(
+            pudl.helpers.organize_cols,
+            [
+                "report_year",
+                "utility_id_ferc1",
+                "utility_id_pudl",
+                "utility_name_ferc1",
+                "plant_name_ferc1",
+                "record_id",
+            ],
+        )
+    )
+    return pumped_storage_df
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_fuel_ferc1(
+    fuel_ferc1: pd.DataFrame, denorm_plants_utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe related to FERC Form 1 fuel information.
+
+    This function pulls the FERC Form 1 fuel data, and joins in the name of the
+    reporting utility, as well as the PUDL IDs for that utility and the plant, allowing
+    integration with other PUDL tables. Useful derived values include:
+
+    * ``fuel_consumed_mmbtu`` (total fuel heat content consumed)
+    * ``fuel_consumed_total_cost`` (total cost of that fuel)
+
+    Args:
+        pudl_engine (sqlalchemy.engine.Engine): Engine for connecting to the
+            PUDL database.
+
+    Returns:
+        A DataFrame containing useful FERC Form 1 fuel
+        information.
+    """
+    fuel_df = (
+        fuel_ferc1.assign(
+            fuel_consumed_mmbtu=lambda x: x["fuel_consumed_units"]
+            * x["fuel_mmbtu_per_unit"],
+            fuel_consumed_total_cost=lambda x: x["fuel_consumed_units"]
+            * x["fuel_cost_per_unit_burned"],
+        )
+        .merge(
+            denorm_plants_utilities_ferc1,
+            on=["utility_id_ferc1", "plant_name_ferc1"],
+        )
+        .pipe(
+            pudl.helpers.organize_cols,
+            [
+                "report_year",
+                "utility_id_ferc1",
+                "utility_id_pudl",
+                "utility_name_ferc1",
+                "plant_id_pudl",
+                "plant_name_ferc1",
+            ],
+        )
+    )
+    return fuel_df
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_purchased_power_ferc1(
+    purchased_power_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    purchased_power_df = purchased_power_ferc1.merge(
+        utilities_ferc1, on="utility_id_ferc1"
+    ).pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+            "seller_name",
+            "record_id",
+        ],
+    )
+    return purchased_power_df
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_plant_in_service_ferc1(
+    plant_in_service_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a dataframe of FERC Form 1 Electric Plant in Service data."""
+    pis_df = plant_in_service_ferc1.merge(utilities_ferc1, on="utility_id_ferc1").pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+            "record_id",
+        ],
+    )
+    return pis_df
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_balance_sheet_assets_ferc1(
+    balance_sheet_assets_ferc1: pd.DataFrame,
+    utilities_ferc1: pd.DataFrame,
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 balance sheet assets data."""
+    denorm_balance_sheet_assets_ferc1 = balance_sheet_assets_ferc1.merge(
+        utilities_ferc1, on="utility_id_ferc1"
+    ).pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+            "record_id",
+            "asset_type",
+        ],
+    )
+    return denorm_balance_sheet_assets_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_balance_sheet_liabilities_ferc1(
+    balance_sheet_liabilities_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 balance_sheet liabilities data."""
+    denorm_balance_sheet_liabilities_ferc1 = balance_sheet_liabilities_ferc1.merge(
+        utilities_ferc1, on="utility_id_ferc1"
+    ).pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+            "record_id",
+            "liability_type",
+        ],
+    )
+    return denorm_balance_sheet_liabilities_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_cash_flow_ferc1(
+    cash_flow_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 cash flow data."""
+    denorm_cash_flow_ferc1 = cash_flow_ferc1.merge(
+        utilities_ferc1, on="utility_id_ferc1"
+    ).pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+            "record_id",
+            "amount_type",
+        ],
+    )
+    return denorm_cash_flow_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_depreciation_amortization_summary_ferc1(
+    depreciation_amortization_summary_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 depreciation amortization data."""
+    denorm_depreciation_amortization_summary_ferc1 = (
+        depreciation_amortization_summary_ferc1.merge(
+            utilities_ferc1, on="utility_id_ferc1"
+        ).pipe(
+            pudl.helpers.organize_cols,
+            [
+                "report_year",
+                "utility_id_ferc1",
+                "utility_id_pudl",
+                "utility_name_ferc1",
+                "record_id",
+                "plant_function",
+                "ferc_account_label",
+            ],
+        )
+    )
+    return denorm_depreciation_amortization_summary_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_electric_energy_dispositions_ferc1(
+    electric_energy_dispositions_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 energy dispositions data."""
+    denorm_electric_energy_dispositions_ferc1 = (
+        electric_energy_dispositions_ferc1.merge(
+            utilities_ferc1, on="utility_id_ferc1"
+        ).pipe(
+            pudl.helpers.organize_cols,
+            [
+                "report_year",
+                "utility_id_ferc1",
+                "utility_id_pudl",
+                "utility_name_ferc1",
+                "record_id",
+                "energy_disposition_type",
+            ],
+        )
+    )
+    return denorm_electric_energy_dispositions_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_electric_energy_sources_ferc1(
+    electric_energy_sources_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    denorm_electric_energy_sources_ferc1 = electric_energy_sources_ferc1.merge(
+        utilities_ferc1, on="utility_id_ferc1"
+    ).pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+            "record_id",
+            "energy_source_type",
+        ],
+    )
+    return denorm_electric_energy_sources_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_electric_operating_expenses_ferc1(
+    electric_operating_expenses_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    denorm_electric_operating_expenses_ferc1 = electric_operating_expenses_ferc1.merge(
+        utilities_ferc1, on="utility_id_ferc1"
+    ).pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+            "record_id",
+            "expense_type",
+        ],
+    )
+    return denorm_electric_operating_expenses_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_electric_operating_revenues_ferc1(
+    electric_operating_revenues_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    denorm_electric_operating_revenues_ferc1 = electric_operating_revenues_ferc1.merge(
+        utilities_ferc1, on="utility_id_ferc1"
+    ).pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+            "record_id",
+            "revenue_type",
+        ],
+    )
+    return denorm_electric_operating_revenues_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_electric_plant_depreciation_changes_ferc1(
+    electric_plant_depreciation_changes_ferc1: pd.DataFrame,
+    utilities_ferc1: pd.DataFrame,
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    denorm_electric_plant_depreciation_changes_ferc1 = (
+        electric_plant_depreciation_changes_ferc1.merge(
+            utilities_ferc1, on="utility_id_ferc1"
+        ).pipe(
+            pudl.helpers.organize_cols,
+            [
+                "report_year",
+                "utility_id_ferc1",
+                "utility_id_pudl",
+                "utility_name_ferc1",
+                "record_id",
+                "depreciation_type",
+                "plant_status",
+                "utility_type",
+            ],
+        )
+    )
+    return denorm_electric_plant_depreciation_changes_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_electric_plant_depreciation_functional_ferc1(
+    electric_plant_depreciation_functional_ferc1: pd.DataFrame,
+    utilities_ferc1: pd.DataFrame,
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    denorm_electric_plant_depreciation_functional_ferc1 = (
+        electric_plant_depreciation_functional_ferc1.merge(
+            utilities_ferc1, on="utility_id_ferc1"
+        ).pipe(
+            pudl.helpers.organize_cols,
+            [
+                "report_year",
+                "utility_id_ferc1",
+                "utility_id_pudl",
+                "utility_name_ferc1",
+                "record_id",
+                "plant_function",
+                "plant_status",
+                "utility_type",
+            ],
+        )
+    )
+    return denorm_electric_plant_depreciation_functional_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_electricity_sales_by_rate_schedule_ferc1(
+    electricity_sales_by_rate_schedule_ferc1: pd.DataFrame,
+    utilities_ferc1: pd.DataFrame,
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    denorm_electricity_sales_by_rate_schedule_ferc1 = (
+        electricity_sales_by_rate_schedule_ferc1.merge(
+            utilities_ferc1, on="utility_id_ferc1"
+        ).pipe(
+            pudl.helpers.organize_cols,
+            [
+                "report_year",
+                "utility_id_ferc1",
+                "utility_id_pudl",
+                "utility_name_ferc1",
+                "record_id",
+            ],
+        )
+    )
+    return denorm_electricity_sales_by_rate_schedule_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_income_statement_ferc1(
+    income_statement_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    denorm_income_statement_ferc1 = income_statement_ferc1.merge(
+        utilities_ferc1, on="utility_id_ferc1"
+    ).pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+            "record_id",
+            "utility_type",
+            "income_type",
+        ],
+    )
+    return denorm_income_statement_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_other_regulatory_liabilities_ferc1(
+    other_regulatory_liabilities_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    denorm_other_regulatory_liabilities_ferc1 = (
+        other_regulatory_liabilities_ferc1.merge(
+            utilities_ferc1, on="utility_id_ferc1"
+        ).pipe(
+            pudl.helpers.organize_cols,
+            [
+                "report_year",
+                "utility_id_ferc1",
+                "utility_id_pudl",
+                "utility_name_ferc1",
+            ],
+        )
+    )
+    return denorm_other_regulatory_liabilities_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_retained_earnings_ferc1(
+    retained_earnings_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    denorm_retained_earnings_ferc1 = retained_earnings_ferc1.merge(
+        utilities_ferc1, on="utility_id_ferc1"
+    ).pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+            "record_id",
+            "earnings_type",
+        ],
+    )
+    return denorm_retained_earnings_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_transmission_statistics_ferc1(
+    transmission_statistics_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    denorm_transmission_statistics_ferc1 = transmission_statistics_ferc1.merge(
+        utilities_ferc1, on="utility_id_ferc1"
+    ).pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+        ],
+    )
+    return denorm_transmission_statistics_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_utility_plant_summary_ferc1(
+    utility_plant_summary_ferc1: pd.DataFrame, utilities_ferc1: pd.DataFrame
+) -> pd.DataFrame:
+    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
+    denorm_utility_plant_summary_ferc1 = utility_plant_summary_ferc1.merge(
+        utilities_ferc1, on="utility_id_ferc1"
+    ).pipe(
+        pudl.helpers.organize_cols,
+        [
+            "report_year",
+            "utility_id_ferc1",
+            "utility_id_pudl",
+            "utility_name_ferc1",
+            "record_id",
+            "utility_type",
+            "utility_plant_asset_type",
+        ],
+    )
+    return denorm_utility_plant_summary_ferc1
+
+
+@asset(io_manager_key="pudl_sqlite_io_manager", compute_kind="Python")
+def denorm_plants_all_ferc1(
+    denorm_plants_steam_ferc1: pd.DataFrame,
+    denorm_plants_small_ferc1: pd.DataFrame,
+    denorm_plants_hydro_ferc1: pd.DataFrame,
+    denorm_plants_pumped_storage_ferc1: pd.DataFrame,
+) -> pd.DataFrame:
+    """Combine the steam, small generators, hydro, and pumped storage tables.
+
+    While this table may have many purposes, the main one is to prepare it for
+    integration with the EIA Master Unit List (MUL). All subtables included in this
+    output table must have pudl ids. Table prepping involves ensuring that the
+    individual tables can merge correctly (like columns have the same name) both with
+    each other and the EIA MUL.
+    """
+    # Prep steam table
+    logger.debug("prepping steam table")
+    steam_df = denorm_plants_steam_ferc1.rename(columns={"opex_plants": "opex_plant"})
+
+    # Prep hydro tables (Add this to the meta data later)
+    logger.debug("prepping hydro tables")
+    hydro_df = denorm_plants_hydro_ferc1.rename(
+        columns={"project_num": "ferc_license_id"}
+    )
+    pump_df = denorm_plants_pumped_storage_ferc1.rename(
+        columns={"project_num": "ferc_license_id"}
+    )
+
+    # Combine all the tables together
+    logger.debug("combining all tables")
+    all_df = (
+        pd.concat([steam_df, denorm_plants_small_ferc1, hydro_df, pump_df])
+        .rename(
+            columns={
+                "fuel_cost": "total_fuel_cost",
+                "fuel_mmbtu": "total_mmbtu",
+                "opex_fuel_per_mwh": "fuel_cost_per_mwh",
+                "primary_fuel_by_mmbtu": "fuel_type_code_pudl",
+            }
+        )
+        .replace({"": np.nan})
+    )
+
+    return all_df
+
+
+@asset(
+    io_manager_key="pudl_sqlite_io_manager",
+    config_schema={
+        "thresh": Field(
+            float,
+            default_value=0.5,
+            description=(
+                "Minimum fraction of fuel (cost and mmbtu) required in order for a "
+                "plant to be assigned a primary fuel. Must be between 0.5 and 1.0. "
+                "Default value is 0.5."
+            ),
+        )
+    },
+    compute_kind="Python",
+)
+def denorm_fuel_by_plant_ferc1(
+    context,
+    fuel_ferc1: pd.DataFrame,
+    denorm_plants_utilities_ferc1: pd.DataFrame,
+) -> pd.DataFrame:
+    """Summarize FERC fuel data by plant for output.
+
+    This is mostly a wrapper around
+    :func:`pudl.analysis.classify_plants_ferc1.fuel_by_plant_ferc1`
+    which calculates some summary values on a per-plant basis (as indicated
+    by ``utility_id_ferc1`` and ``plant_name_ferc1``) related to fuel
+    consumption.
+
+    Args:
+        context: Dagster context object
+        fuel_ferc1: Normalized FERC fuel table.
+        denorm_plants_utilities_ferc1: Denormalized table of FERC1 plant & utility IDs.
+
+    Returns:
+        A DataFrame with fuel use summarized by plant.
+    """
+
+    def drop_other_fuel_types(df):
+        """Internal function to drop other fuel type.
+
+        Fuel type other indicates we didn't know how to categorize the reported fuel
+        type, which leads to records with incomplete and unsable data.
+        """
+        return df[df.fuel_type_code_pudl != "other"].copy()
+
+    thresh = context.op_config["thresh"]
+    # The existing function expects `fuel_type_code_pudl` to be an object, rather than
+    # a category. This is a legacy of pre-dagster code, and we convert here to prevent
+    # further retooling in the code-base.
+    fuel_ferc1["fuel_type_code_pudl"] = fuel_ferc1["fuel_type_code_pudl"].astype(str)
+
+    fuel_categories = list(
+        pudl.transform.ferc1.FuelFerc1TableTransformer()
+        .params.categorize_strings["fuel_type_code_pudl"]
+        .categories.keys()
+    )
+
+    fbp_df = (
+        fuel_ferc1.pipe(drop_other_fuel_types)
+        .pipe(
+            pudl.analysis.classify_plants_ferc1.fuel_by_plant_ferc1,
+            fuel_categories=fuel_categories,
+            thresh=thresh,
+        )
+        .pipe(pudl.analysis.classify_plants_ferc1.revert_filled_in_float_nulls)
+        .pipe(pudl.analysis.classify_plants_ferc1.revert_filled_in_string_nulls)
+        .merge(
+            denorm_plants_utilities_ferc1, on=["utility_id_ferc1", "plant_name_ferc1"]
+        )
+        .pipe(
+            pudl.helpers.organize_cols,
+            [
+                "report_year",
+                "utility_id_ferc1",
+                "utility_id_pudl",
+                "utility_name_ferc1",
+                "plant_id_pudl",
+                "plant_name_ferc1",
+            ],
+        )
+    )
+    return fbp_df
+
+
+###########################################################################
+# HELPER FUNCTIONS
+###########################################################################
 
 
 def calc_annual_capital_additions_ferc1(
@@ -243,291 +871,3 @@ def add_mean_cap_additions(steam_df):
         )
     )
     return df
-
-
-def fuel_ferc1(pudl_engine, start_date, end_date):
-    """Pull a useful dataframe related to FERC Form 1 fuel information.
-
-    This function pulls the FERC Form 1 fuel data, and joins in the name of the
-    reporting utility, as well as the PUDL IDs for that utility and the plant,
-    allowing integration with other PUDL tables.
-
-    Useful derived values include:
-
-    * ``fuel_consumed_mmbtu`` (total fuel heat content consumed)
-    * ``fuel_consumed_total_cost`` (total cost of that fuel)
-
-    Args:
-        pudl_engine (sqlalchemy.engine.Engine): Engine for connecting to the
-            PUDL database.
-
-    Returns:
-        pandas.DataFrame: A DataFrame containing useful FERC Form 1 fuel
-        information.
-    """
-    fuel_df = (
-        read_table_with_start_end_dates("fuel_ferc1", start_date, end_date, pudl_engine)
-        .assign(
-            fuel_consumed_mmbtu=lambda x: x["fuel_consumed_units"]
-            * x["fuel_mmbtu_per_unit"],
-            fuel_consumed_total_cost=lambda x: x["fuel_consumed_units"]
-            * x["fuel_cost_per_unit_burned"],
-        )
-        .merge(
-            plants_utils_ferc1(pudl_engine), on=["utility_id_ferc1", "plant_name_ferc1"]
-        )
-        .pipe(
-            pudl.helpers.organize_cols,
-            [
-                "report_year",
-                "utility_id_ferc1",
-                "utility_id_pudl",
-                "utility_name_ferc1",
-                "plant_id_pudl",
-                "plant_name_ferc1",
-            ],
-        )
-    )
-    return fuel_df
-
-
-def fuel_by_plant_ferc1(pudl_engine, start_date, end_date, thresh=0.5):
-    """Summarize FERC fuel data by plant for output.
-
-    This is mostly a wrapper around
-    :func:`pudl.analysis.classify_plants_ferc1.fuel_by_plant_ferc1`
-    which calculates some summary values on a per-plant basis (as indicated
-    by ``utility_id_ferc1`` and ``plant_name_ferc1``) related to fuel
-    consumption.
-
-    Args:
-        pudl_engine (sqlalchemy.engine.Engine): Engine for connecting to the
-            PUDL database.
-        thresh (float): Minimum fraction of fuel (cost and mmbtu) required in
-            order for a plant to be assigned a primary fuel. Must be between
-            0.5 and 1.0. default value is 0.5.
-
-    Returns:
-        pandas.DataFrame: A DataFrame with fuel use summarized by plant.
-    """
-
-    def drop_other_fuel_types(df):
-        """Internal function to drop other fuel type.
-
-        Fuel type other indicates we didn't know how to categorize the reported fuel
-        type, which leads to records with incomplete and unsable data.
-        """
-        return df[df.fuel_type_code_pudl != "other"].copy()
-
-    fuel_categories = list(
-        pudl.transform.ferc1.FuelFerc1TableTransformer()
-        .params.categorize_strings["fuel_type_code_pudl"]
-        .categories.keys()
-    )
-    fbp_df = (
-        read_table_with_start_end_dates("fuel_ferc1", start_date, end_date, pudl_engine)
-        .pipe(drop_other_fuel_types)
-        .pipe(
-            pudl.analysis.classify_plants_ferc1.fuel_by_plant_ferc1,
-            fuel_categories=fuel_categories,
-            thresh=thresh,
-        )
-        .pipe(pudl.analysis.classify_plants_ferc1.revert_filled_in_float_nulls)
-        .pipe(pudl.analysis.classify_plants_ferc1.revert_filled_in_string_nulls)
-        .merge(
-            plants_utils_ferc1(pudl_engine), on=["utility_id_ferc1", "plant_name_ferc1"]
-        )
-        .pipe(
-            pudl.helpers.organize_cols,
-            [
-                "report_year",
-                "utility_id_ferc1",
-                "utility_id_pudl",
-                "utility_name_ferc1",
-                "plant_id_pudl",
-                "plant_name_ferc1",
-            ],
-        )
-    )
-    return fbp_df
-
-
-def plants_small_ferc1(pudl_engine, start_date, end_date):
-    """Pull a useful dataframe related to the FERC Form 1 small plants."""
-    plants_small_df = (
-        read_table_with_start_end_dates(
-            "plants_small_ferc1", start_date, end_date, pudl_engine
-        )
-        .merge(
-            plants_utils_ferc1(pudl_engine),
-            on=["utility_id_ferc1", "plant_name_ferc1"],
-            how="left",
-        )
-        .assign(
-            opex_total=lambda x: (
-                x[["opex_fuel", "opex_maintenance", "opex_operations"]]
-                .fillna(0)
-                .sum(axis=1)
-            ),
-            opex_total_nonfuel=lambda x: (x.opex_total - x.opex_fuel.fillna(0)),
-        )
-        .pipe(
-            pudl.helpers.organize_cols,
-            [
-                "report_year",
-                "utility_id_ferc1",
-                "utility_id_pudl",
-                "utility_name_ferc1",
-                "plant_id_pudl",
-                "plant_name_ferc1",
-                "record_id",
-            ],
-        )
-    )
-    return plants_small_df
-
-
-def plants_hydro_ferc1(pudl_engine, start_date, end_date):
-    """Pull a useful dataframe related to the FERC Form 1 hydro plants."""
-    plants_hydro_df = (
-        read_table_with_start_end_dates(
-            "plants_hydro_ferc1", start_date, end_date, pudl_engine
-        )
-        .merge(
-            plants_utils_ferc1(pudl_engine),
-            on=["utility_id_ferc1", "plant_name_ferc1"],
-            how="left",
-        )
-        .assign(
-            capacity_factor=lambda x: (x.net_generation_mwh / (8760 * x.capacity_mw)),
-            opex_total_nonfuel=lambda x: x.opex_total,
-        )
-        .pipe(
-            pudl.helpers.organize_cols,
-            [
-                "report_year",
-                "utility_id_ferc1",
-                "utility_id_pudl",
-                "utility_name_ferc1",
-                "plant_name_ferc1",
-                "record_id",
-            ],
-        )
-    )
-    return plants_hydro_df
-
-
-def plants_pumped_storage_ferc1(pudl_engine, start_date, end_date):
-    """Pull a dataframe of FERC Form 1 Pumped Storage plant data."""
-    pumped_storage_df = (
-        read_table_with_start_end_dates(
-            "plants_pumped_storage_ferc1", start_date, end_date, pudl_engine
-        )
-        .merge(
-            pudl.output.ferc1.plants_utils_ferc1(pudl_engine),
-            on=["utility_id_ferc1", "plant_name_ferc1"],
-            how="left",
-        )
-        .assign(
-            capacity_factor=lambda x: x.net_generation_mwh / (8760 * x.capacity_mw),
-            opex_total_nonfuel=lambda x: x.opex_total,
-        )
-        .pipe(
-            pudl.helpers.organize_cols,
-            [
-                "report_year",
-                "utility_id_ferc1",
-                "utility_id_pudl",
-                "utility_name_ferc1",
-                "plant_name_ferc1",
-                "record_id",
-            ],
-        )
-    )
-    return pumped_storage_df
-
-
-def purchased_power_ferc1(pudl_engine, start_date, end_date):
-    """Pull a useful dataframe of FERC Form 1 Purchased Power data."""
-    purchased_power_df = (
-        read_table_with_start_end_dates(
-            "purchased_power_ferc1", start_date, end_date, pudl_engine
-        )
-        .merge(pd.read_sql("utilities_ferc1", pudl_engine), on="utility_id_ferc1")
-        .pipe(
-            pudl.helpers.organize_cols,
-            [
-                "report_year",
-                "utility_id_ferc1",
-                "utility_id_pudl",
-                "utility_name_ferc1",
-                "seller_name",
-                "record_id",
-            ],
-        )
-    )
-    return purchased_power_df
-
-
-def plant_in_service_ferc1(pudl_engine, start_date, end_date):
-    """Pull a dataframe of FERC Form 1 Electric Plant in Service data."""
-    pis_df = (
-        read_table_with_start_end_dates(
-            "plant_in_service_ferc1", start_date, end_date, pudl_engine
-        )
-        .merge(pd.read_sql("utilities_ferc1", pudl_engine), on="utility_id_ferc1")
-        .pipe(
-            pudl.helpers.organize_cols,
-            [
-                "report_year",
-                "utility_id_ferc1",
-                "utility_id_pudl",
-                "utility_name_ferc1",
-                "record_id",
-            ],
-        )
-    )
-    return pis_df
-
-
-def plants_all_ferc1(pudl_engine, start_date, end_date):
-    """Combine the steam, small generators, hydro, and pumped storage tables.
-
-    While this table may have many purposes, the main one is to prepare it for
-    integration with the EIA Master Unit List (MUL). All subtables included in this
-    output table must have pudl ids. Table prepping involves ensuring that the
-    individual tables can merge correctly (like columns have the same name) both with
-    each other and the EIA MUL.
-    """
-    steam_df = plants_steam_ferc1(pudl_engine, start_date, end_date)
-    small_df = plants_small_ferc1(pudl_engine, start_date, end_date)
-    hydro_df = plants_hydro_ferc1(pudl_engine, start_date, end_date)
-    pump_df = plants_pumped_storage_ferc1(pudl_engine, start_date, end_date)
-
-    # Prep steam table
-    logger.debug("prepping steam table")
-    steam_df = steam_df.rename(columns={"opex_plants": "opex_plant"}).pipe(
-        apply_pudl_dtypes, group="ferc1"
-    )
-
-    # Prep hydro tables (Add this to the meta data later)
-    logger.debug("prepping hydro tables")
-    hydro_df = hydro_df.rename(columns={"project_num": "ferc_license_id"})
-    pump_df = pump_df.rename(columns={"project_num": "ferc_license_id"})
-
-    # Combine all the tables together
-    logger.debug("combining all tables")
-    all_df = (
-        pd.concat([steam_df, small_df, hydro_df, pump_df])
-        .rename(
-            columns={
-                "fuel_cost": "total_fuel_cost",
-                "fuel_mmbtu": "total_mmbtu",
-                "opex_fuel_per_mwh": "fuel_cost_per_mwh",
-                "primary_fuel_by_mmbtu": "fuel_type_code_pudl",
-            }
-        )
-        .replace({"": np.nan})
-    )
-
-    return all_df
